requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.6.3
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blessings==1.7
cachetools==4.2.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.7
Django==5.0.6
django-cors-headers==4.3.1
django-geoposition==0.3.0
django-redis==5.4.0
djangorestframework==3.15.1
earthengine-api==0.1.405
ee==0.2
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googlemaps==4.10.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
multidict==6.0.5
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-telegram-bot==13.13
pytz==2024.1
razorpay==1.4.2
redis==5.0.4
requests==2.32.3
retrying==1.3.4
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tornado==6.4
twilio==9.1.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
zope.interface==6.4.post2
